Pong: Remove WVCdm L1 vendor lib
Quite sadly, L1 breaks as soon as bootloader is unlocked.
Despite apps such as DRM Info still report security level 1,
it can be noticed that system ID is not proper as it shows
a 10 digit number instead of 5. In this state, the device
doesn't properly fallback to L3, thus breaking any app that
requires DRM.

Removing the lib allows the sucessful fallback to L3,
at least it's something...

Change-Id: I8e09647674cdb98a85a4977b1b468288c4ca52c5
Signed-off-by: chandu078 <chandudyavanapelli03@gmail.com>